backend/requirements.txt
argon2-cffi==23.1.0
channels==4.0.0
coreapi==2.3.3
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-tinymce==4.1.0
gunicorn==21.2.0
mutagen==1.47.0
Pillow==10.1.0
pip-review==1.3.0  
pipreqs==0.4.13
psycopg2-binary==2.9.9
requests==2.31.0
whitenoise==6.6.0
pipdeptree==2.13.1
    

#setuptools==68.2.2
# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
